Trapped02 - now with 4 macros AND retriggering
The idea here is that you grab a 'fruit' with the "Grip" button and then, if you are holding in on the 'Trigger" button you can use macros 1,2,3 to change the sound.  AND, is you hold in "Button A" on the Right Controller or "Button X" on the left controller (if you have grabbed a 'fruit' with the left controller, then ROTATION sets the retrigger rate.  When you release - the settings stay where you left them(?) You can grab 2 fruits at a time - one with each hand.  For both  - macro1 = pitch on the Y axis, macro2 = filter on the X axis, macro3 = volume & reverb depth - (softer has more reverb) on the Z axis, And retrigger rates are set with rotation.